Lab-2/requirements.txt
python 3.12
pytest
python-docx
platform
os
abc